% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_prism.R
\name{aggregate_prism}
\alias{aggregate_prism}
\title{Generalized version of PRISM_ym
Takes in a data.table with daily weather data and produces aggregated data.table by month of sample using requested summary functions.}
\usage{
aggregate_prism(dt, i.names, t.in = c("date"), t.fun = as.yearmon,
  mean.names = c("tmin"), sum.names = c("ppt"), bins.seqs = list(tmax =
  c(-Inf, seq(0, 40, 10), Inf), tmin = c(-Inf, seq(-10, 30, 10), Inf)))
}
\arguments{
\item{dt}{data.table to summarize.}

\item{i.names}{Character vector of variables in dt.file that identify}

\item{t.in}{Character vector of length 1 that identifies input time variable.}

\item{t.fun}{Single item list, name 1 is grouping name and item 1 is grouping function to execute on t.in}

\item{mean.names}{Character vector of variables to take mean over.}

\item{sum.names}{Character vector of variables to sum over.}

\item{bins.seqs}{List of named sequences that define bin variables.}
}
\value{
Monthly summary of variables from PRISM.
}
\description{
Generalized version of PRISM_ym
Takes in a data.table with daily weather data and produces aggregated data.table by month of sample using requested summary functions.
}
