% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_prism_ym.R
\name{aggregate_prism_ym}
\alias{aggregate_prism_ym}
\title{Aggregate PRISM file by ym}
\usage{
aggregate_prism_ym(dt.file, i.names, t.name, bincount.seqs = NULL,
  mean.names = NULL, sum.names = NULL)
}
\arguments{
\item{dt.file}{filename of .Rds with data.table}

\item{i.names}{Name(s) of ID variable}

\item{t.name}{Name of time variable}

\item{bincount.seqs}{Named list of sequences for cut variables. Name should be the relevant variable name.}

\item{mean.names}{Character vector of names of variables to take means over}

\item{sum.names}{Character vector of names of variables to take sums over}
}
\value{
data.table of aggregated data
}
\description{
Aggregate PRISM file by ym
}
