% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bound_var.R
\name{bound_var}
\alias{bound_var}
\title{Constrain values of a numeric vector within lower and upper limits}
\usage{
bound_var(x, x_min = 0, x_max = 30000)
}
\arguments{
\item{x}{Numeric vector}

\item{x_min}{Minimum allowable value of x}

\item{x_max}{Maximum allowable value of x}
}
\value{
Returns numeric vector within all values of x  x_min and x_max
}
\description{
Constrain values of a numeric vector within lower and upper limits
}
