% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_binned_results.R
\name{clean_binned_results}
\alias{clean_binned_results}
\title{Clean binned results from felm (and maybe lm?). Will try to guess at omitted variable from variable names created using cut().}
\usage{
clean_binned_results(felm.est = NULL, felm.dt = NULL,
  keepvars = NULL)
}
\arguments{
\item{felm.est}{felm output from regression with binned variables, can be stripped with strip_felm. Required if felm.dt not supplied.}

\item{felm.dt}{data.table of output from regression with binned variables. Required if felm.est not supplied.}

\item{keepvars}{(Optional) regex of variables to keep.}
}
\value{
Ordered data.table with coef, se, ci.l, ci.h, xmin, xmax, xmid. Will include a row for the (guessed at) omitted variable with coef and se = 0.
}
\description{
Clean binned results from felm (and maybe lm?). Will try to guess at omitted variable from variable names created using cut().
}
