% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut_and_reorder.R
\name{cut_and_reorder}
\alias{cut_and_reorder}
\title{Reorder a variable quickly for a regression.}
\usage{
cut_and_reorder(x, cut.seq = c(-Inf, seq(0, 40, 5), Inf), omit = 22.5)
}
\arguments{
\item{x}{Numeric vector.}

\item{cut.seq}{(Optional) Numeric vector to pass to cut}

\item{omit}{(Optional) Bin with this number will be placed first in vector.}
}
\value{
Factorized version of x, with given sequence and omitted level first.
}
\description{
Reorder a variable quickly for a regression.
}
