% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duplicated_all.R
\name{duplicated_all}
\alias{duplicated_all}
\title{Identify rows with duplicates}
\usage{
duplicated_all(x, by)
}
\arguments{
\item{x}{data.frame or data.table}

\item{by}{character vector of variables over which to check duplicates}
}
\value{
boolean vector, where TRUE indicates that at least one other row in x has the same values for by
}
\description{
Identify rows with duplicates
}
