% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_binned_results.R
\name{extract_from_range}
\alias{extract_from_range}
\title{Get varname, min, max from named range. Not exported}
\usage{
extract_from_range(x)
}
\arguments{
\item{x}{Character vector.}
}
\value{
Named list with varname, xmin, xmax
}
\description{
Get varname, min, max from named range. Not exported
}
