% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stargazer_functions.R
\name{fC}
\alias{fC}
\title{Convenience function to format number with commas.}
\usage{
fC(x, digits = 2)
}
\arguments{
\item{x}{Numeric vector.}

\item{digits}{Number of digits past decimal.}
}
\value{
Character vector with commas.
}
\description{
Convenience function to format number with commas.
}
