% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_binned_results.R
\name{felm_to_dt}
\alias{felm_to_dt}
\title{# Convert felm to data.table. Not exported for now.}
\usage{
felm_to_dt(felm.est, keepvars = NULL)
}
\arguments{
\item{felm.est}{felm object.}

\item{keepvars}{regex of variables to keep.}
}
\value{
data.table of results
}
\description{
# Convert felm to data.table. Not exported for now.
}
