% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_binned_results.R
\name{find_missing}
\alias{find_missing}
\title{# By varname, fill in missing value, assuming constant gap. Used by clean_binned_results, not for export right now.}
\usage{
find_missing(x, gap)
}
\arguments{
\item{x}{Numeric vector}

\item{gap}{Numeric of gap between each x.}
}
\value{
Numeric of missing x value.
}
\description{
# By varname, fill in missing value, assuming constant gap. Used by clean_binned_results, not for export right now.
}
