% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_dups.R
\name{get_all_dups}
\alias{get_all_dups}
\title{Get all duplicates in a data.table for a given set of variables}
\usage{
get_all_dups(dt, by)
}
\arguments{
\item{dt}{data.table}

\item{by}{character vector of variable names to use as ID columns for duplicate check}
}
\value{
Returns data table of duplicates
}
\description{
Get all duplicates in a data.table for a given set of variables
}
