% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_twitter_files.R
\name{get_local_time}
\alias{get_local_time}
\title{Gets local time for a data.table with lat, lng, datetimeUTC}
\usage{
get_local_time(dt, debug = FALSE)
}
\arguments{
\item{dt}{data.table to import, must have following cols (at minimum): lng, lat, tweet_datetime.}

\item{debug}{Flag to also return raster cell, value, and TZID.
#' @return Returns data.table.}
}
\description{
Gets local time for a data.table with lat, lng, datetimeUTC
}
