% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_twitter_csv.R
\name{import_twitter_csv}
\alias{import_twitter_csv}
\title{This function imports a single CSV of twitter data.}
\usage{
import_twitter_csv(inFile, col_types = NULL, rast = NULL,
  keep_UTC = FALSE)
}
\arguments{
\item{inFile}{Filename to import (.csv or .csv.gz).}

\item{col_types}{Passed through to read_csv. NULL imports all fields.}

\item{keep_UTC}{Flag to return UTC datetime.}
}
\value{
Returns data.table.
import_twitter_csv()
}
\description{
This function imports a single CSV of twitter data.
}
