% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_twitter_files.R
\name{import_twitter_file}
\alias{import_twitter_file}
\title{This function imports a single CSV of twitter data.}
\usage{
import_twitter_file(inFile, col_types = NULL)
}
\arguments{
\item{inFile}{Filename to import (.csv or .csv.gz).}

\item{col_types}{Passed through to read_csv. NULL imports all fields.}
}
\value{
Returns data.table.
}
\description{
This function imports a single CSV of twitter data.
}
