% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_twitter_csv.R
\name{import_twitter_files}
\alias{import_twitter_files}
\title{This function imports many CSVs of twitter data and collects them into a single data.table.}
\usage{
import_twitter_files(inFiles, col_types = NULL, rast.file = NULL,
  keep_UTC = FALSE)
}
\arguments{
\item{inFiles}{Set of files to import. Presumably a list of .csv or .csv.gz.}

\item{col_types}{Passed through to read_csv. NULL imports all fields.}

\item{keep_UTC}{Flag to return UTC datetime.}
}
\value{
Returns data.table.
import_twitter_dir()
}
\description{
This function imports many CSVs of twitter data and collects them into a single data.table.
}
