% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stargazer_functions.R
\name{latex_makesubrules}
\alias{latex_makesubrules}
\title{Create a set of cmidrules based on given edges.}
\usage{
latex_makesubrules(edges)
}
\arguments{
\item{edges}{Should be an even number.}
}
\value{
LaTeX character vector of cmidrules.
}
\description{
Create a set of cmidrules based on given edges.
}
