% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stargazer_functions.R
\name{latex_special_row}
\alias{latex_special_row}
\title{Create a LaTeX table row character vector wrapped in multicolumn.}
\usage{
latex_special_row(x, col.widths = NULL, align = NULL)
}
\arguments{
\item{x}{vector with length = to number of columns, will be coerced to character}

\item{col.widths}{Vector of column widths. Will be assumed to be 1 for member of x if not specified.}

\item{align}{Desired alignment for cells.}
}
\value{
LaTeX table row character vector wrapped in multicolumn.
}
\description{
Create a LaTeX table row character vector wrapped in multicolumn.
}
