% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obj_size.R
\name{obj_size}
\alias{obj_size}
\title{Return data.table with lag}
\usage{
obj_size(x)
}
\arguments{
\item{x}{Name of object}
}
\value{
Prints string with x's size, units chosen automatically
}
\description{
Return data.table with lag
}
