% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pastep.R
\name{pastep}
\alias{pastep}
\title{Concatenate and flattens strings with " + "}
\usage{
pastep(...)
}
\arguments{
\item{...}{Character strings to concatenate with " + "}
}
\value{
Returns character vector with " + " in between each
}
\description{
Concatenate and flattens strings with " + "
}
