% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prism_daily_to_ym.R
\name{prism_daily_to_ym}
\alias{prism_daily_to_ym}
\title{Takes in a data.table with daily weather data and produces aggregated data.table by month of sample using requested summary functions.}
\usage{
prism_daily_to_ym(dt, i.names, t.name = c("date"), mean.names = c("tmin"),
  sum.names = c("ppt"), bins.seqs = list(tmax = c(-Inf, seq(0, 40, 10),
  Inf), tmin = c(-Inf, seq(-10, 30, 10), Inf)))
}
\arguments{
\item{dt}{data.table to summarize.}

\item{i.names}{Character vector of variables in dt.file that identify}

\item{t.name}{Character vector of length 1 that identifies date variable.}

\item{mean.names}{Character vector of variables to take mean over.}

\item{sum.names}{Character vector of variables to sum over.}

\item{bins.seqs}{List of named sequences that define bin variables.}
}
\value{
Monthly summary of variables from PRISM.
}
\description{
Takes in a data.table with daily weather data and produces aggregated data.table by month of sample using requested summary functions.
}
