% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg_felm.R
\name{reg_felm}
\alias{reg_felm}
\title{Wrapper for lfe::felm}
\usage{
reg_felm(lhs, rhs, fe, clus, data, weights_var)
}
\arguments{
\item{lhs}{Character string of LHS, e.g., "y"}

\item{rhs}{Character string of RHS, e.g., "x1 + x2"}

\item{fe}{Character string of RHS, e.g., "fe1 + fe2"}

\item{clus}{Character string of clustering variables, e.g., "clus1 + clus2"}

\item{data}{Symbol for data.table with data in it}

\item{weights_var}{Character string for the weights_var (should be found in data)}
}
\value{
Stripped felm object.
}
\description{
Wrapper for lfe::felm
}
