% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stargazer_functions.R
\name{stargazer_clean}
\alias{stargazer_clean}
\title{Clean up stargazer output}
\usage{
stargazer_clean(stargazer.out, replace.stars.with.sym = F)
}
\arguments{
\item{stargazer.out}{Output from stargazer}

\item{replace.stars.with.sym}{Boolean to replace stars with sym command.}
}
\value{
Cleaned-up table output.
}
\description{
Clean up stargazer output
}
