% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stargazer_functions.R
\name{stargazerw}
\alias{stargazerw}
\title{Calls stargazer, returns cleaned up stargazer output. Sets a few other defaults.}
\usage{
stargazerw(..., align = T, table.layout = "c-t", no.space = T,
  style = "aer", fragment = F, drop.hline = F)
}
\arguments{
\item{...}{Parameters for stargazer}
}
\value{
Character vector of cleaned-up table output.
}
\description{
Calls stargazer, returns cleaned up stargazer output. Sets a few other defaults.
}
