% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strip_felm.R
\name{strip_felm}
\alias{strip_felm}
\title{Strip off data-intensive felm objects.}
\usage{
strip_felm(felm.est)
}
\arguments{
\item{felm.est}{Object from lfe::felm call}
}
\value{
Returns (much smaller) lfe::felm object.
}
\description{
Strip off data-intensive felm objects.
}
