% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_lags_and_leads.R
\name{add_lags_and_leads}
\alias{add_lags_and_leads}
\title{Return data.table with lag}
\usage{
add_lags_and_leads(dt, varnames, tvar = "date", by = NULL,
  typenames = c("lag", "lead"))
}
\arguments{
\item{dt}{data.table}

\item{varnames}{Character vector of variables to lag or lead}

\item{tvar}{Name of time variable. Default is "date".}

\item{by}{(Optional) Name of variable to run lag/lead by}

\item{typenames}{(Optional) Character vector equal to c("lag"), c("lead")", or c("lag", "lead") (default)}
}
\value{
data.table with lags and leads added
}
\description{
Return data.table with lag
}
