% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_prism_ym.R
\name{aggregate_prism_ym}
\alias{aggregate_prism_ym}
\title{Reorder a variable quickly for a regression.}
\usage{
aggregate_prism_ym(dt.file, i.names, t.name, bincount.seqs = NULL,
  mean.names = NULL, sum.names = NULL)
}
\arguments{
\item{x}{Numeric vector.}

\item{cut.seq}{(Optional) Numeric vector to pass to cut}

\item{omit}{(Optional) Bin with this number will be placed first in vector.}
}
\value{
Factorized version of x, with given sequence and omitted level first.
}
\description{
Reorder a variable quickly for a regression.
}
